requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.1.3
ItsDangerous==2.2.0
Blinker==1.9.0
WTForms==3.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-magic==0.4.27
Pillow==11.1.0
redis==5.2.0
email_validator==2.2.0
